+++ src/main/java/racing/domain/ClientInput.java
@@ -1,4 +1,4 @@
-package racing.core;
+package racing.domain;
 
 import java.util.Objects;
