+++ src/main/java/racing/domain/Mover.java
@@ -1,4 +1,4 @@
-package racing.core;
+package racing.domain;
 
 public class Mover {
