+++ src/test/java/racing/domain/CarsTest.java
@@ -1,4 +1,4 @@
-package racing.core;
+package racing.domain;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
@@ -8,26 +8,8 @@ import static org.assertj.core.api.Assertions.assertThat;
 public class CarsTest {
 
     @Test
-    @DisplayName("정상적으로 내림차순으로 정렬하는지 확인")
-    void testSortDesc() {
-        Car car1 = new Car(new CarName("jason"));
-        Car car2 = new Car(new CarName("ethan"));
-        Car car3 = new Car(new CarName("pat"));
-        Mover mover = new Mover();
-        mover.decideMove(car3, 9);
-        mover.decideMove(car3, 9);
-        mover.decideMove(car1, 7);
-        Cars cars = Cars.makeCars(car1, car2, car3);
-
-//        cars.sortDesc();
-
-        Cars expected = Cars.makeCars(car3, car1, car2);
-        assertThat(cars).isEqualTo(expected);
-    }
-
-    @Test
     @DisplayName("같은 거리값을 가진 자동차 리스트 반환 확인")
-    void testGetCarsWithSameDistance(){
+    void testGetCarsWithSameDistance() {
         Car car1 = new Car(new CarName("jason"));
         Car car2 = new Car(new CarName("ethan"));
         Car car3 = new Car(new CarName("pat"));
@@ -46,7 +28,7 @@ public class CarsTest {
 
     @Test
     @DisplayName("문자열이 주어졌을 때 자동차 리스트 반환 확인")
-    void testMakeCarWithString(){
+    void testMakeCarWithString() {
         String input = "jack,pat,ethan";
 
         Cars result = Cars.makeCars(input);
