+++ src/test/java/racing/domain/MoverTest.java
@@ -1,4 +1,4 @@
-package racing.core;
+package racing.domain;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
