+++ src/test/java/racing/view/InputViewTest.java
@@ -3,7 +3,7 @@ package racing.view;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import racing.core.ClientInput;
+import racing.domain.ClientInput;
 
 import java.io.*;
