+++ src/test/java/racing/view/ResultViewTest.java
@@ -3,10 +3,10 @@ package racing.view;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import racing.core.Car;
-import racing.core.CarName;
-import racing.core.Cars;
-import racing.core.Mover;
+import racing.domain.Car;
+import racing.domain.CarName;
+import racing.domain.Cars;
+import racing.domain.Mover;
 
 import java.io.ByteArrayOutputStream;
 import java.io.PrintStream;
@@ -50,7 +50,7 @@ public class ResultViewTest {
 
     @Test
     @DisplayName("크기가 1보다 큰 WinnerList가 주어졌을 때 정상 출력값 확인")
-    void testPrintedTextWithWinners(){
+    void testPrintedTextWithWinners() {
         Car car1 = new Car(new CarName("jack"));
         Car car2 = new Car(new CarName("jason"));
 
@@ -61,7 +61,7 @@ public class ResultViewTest {
 
     @Test
     @DisplayName("크기가 1인 WinnerList가 주어졌을 때 정상 출력값 확인")
-    void testPrintedTextWithWinner(){
+    void testPrintedTextWithWinner() {
         Car car1 = new Car(new CarName("jack"));
 
         resultView.printWinners(Cars.makeCars(car1));
